+++ ReleaseNotes.md
@@ -2,8 +2,13 @@
 ## Release Notes
 
 
+__Version 1.0.5__
+* Expose Manager Watchdog so it can be used by top level process
+* Watchdog shutdown event add to synch shutdown process
+
 __Version 1.0.4__
 * Watchdog shutdown stops all registered tasks
+
 
 __Version 1.0.0__
 * Initial Release